requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.0
attrs==23.2.0
b2sdk==2.5.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.150
botocore==1.34.150
bs4==0.0.2
cachetools==5.3.3
captcha==0.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==26.0.0
favicon==0.7.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.2
hydralit-components==1.0.10
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
logfury==1.0.1
lxml==5.2.2
mailtrap==2.0.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
mysql-connector-python==9.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
protobuf==4.25.4
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.7
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
st_tailwind==1.0.3
streamlit==1.14.0
streamlit-antd-components==0.3.2
streamlit-authenticator==0.3.2
streamlit-calendar==1.2.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pdf-viewer==0.0.17
streamlit-searchbox==0.1.14
streamlit-shadcn-ui==0.1.18
streamlit-tailwind==0.1.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.1
